requirements.txt
# Основные зависимости
langchain==0.2.2
langchain-community==0.2.3
langchain-openai==0.1.8
chromadb==0.5.0
python-dotenv==1.0.1
openai==1.31.1
tiktoken==0.7.0

# Загрузчики документов
unstructured==0.14.4
pdf2image==1.17.0
pdfminer.six==20231228
python-docx==1.1.0
python-pptx==0.6.23
beautifulsoup4==4.12.3
markdown==3.5.2
nltk==3.8.1
pandas==2.2.1

# Тестирование
pytest==8.0.2
pytest-cov==4.1.0

# AWS Bedrock (опционально)
boto3==1.34.34

# Ollama (опционально)
ollama==0.1.6

# После установки основных зависимостей выполните:
# pip install "unstructured[md]"
# pip install "unstructured[pdf]"